+++ inst/source/R/vim.R
@@ -16,21 +16,22 @@ call_vim <- function(file, title = file) {
 #'
 #' Just a wrapper to \code{\link{file.edit}}, trying to use [g]vim as editor, if
 #' installed.
-#' @param file See \code{\link{file.edit}}.
+#' @param ... See \code{\link{file.edit}}.
 #' @return See \code{\link{file.edit}}.
 #' @export
 #' @family operating system functions
 #' @examples
-#' path <- file.path(tempdir(), "foo.txt")
-#' writeLines(c("abc", "xyz"), con = path)
-#' vim(path)
-vim <- function(file) {
-    if (interactive()) {
-        if (is_installed("vim")) old <- options(editor = call_vim)
-        res <- utils::file.edit(file = file)
-        if (exists("old", inherits =  FALSE)) options(old)
-    } else {
-        res <- cat(readLines(file), sep = "\n")
+#' if (interactive()) {
+#'     path <- file.path(tempdir(), "foo.txt")
+#'     writeLines(c("abc", "xyz"), con = path)
+#'     vim(path)
+#' }
+vim <- function(...) {
+    if (! interactive()) {
+        warning("Editing files non-interactively.")
     }
+    if (is_installed("vim")) old <- options(editor = call_vim)
+    res <- utils::file.edit(...)
+    if (exists("old", inherits =  FALSE)) options(old)
     return(res)
 }